KTO-422: Move TeemakuvaSection to components-directory - TeemakuvaSection isnot specific to koulutus
